display_serv/requirements.txt
pip==22.1.2
python-dotenv==0.20.0
Flask==2.1.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
config==0.5.1
qrcode==7.3.1
Pillow==9.1.1
pyqrcode==1.2.1
python2-secrets==1.0.5
pytest==7.1.2
colorama==0.4.5
cryptography==37.0.4
wheel==0.37.1
PyMySQL==1.0.2
blinker==1.4.0
alembic==1.8.0
setuptools==62.6.0
click==8.1.3
requests==2.28.1
